Move diskio.cpp into 3rd party
- Fixed typo with debug target.
- Fixed typo with commit-tidy target.
- Moved ram_test.cpp into L1 unity tests.
- Add diskio.cpp to lib3p unity files.
- Move fatfs.hpp to root of L3_Application
- Declared CRP in linker script as INVALID to prevent it from being
  used. Removed CRP word from startup.cpp file. This allow developers
  to use their own linker script or CRP words within their projects.
  This also fixes the issue with barebones reporting invalid CRP because
  of heavy optimization.

Resolves #1282
