+++ library/L1_Peripheral/test/unity_test.cpp
@@ -1,4 +1,10 @@
 // =============================================================================
+// L0 Platform Tests
+// =============================================================================
+
+#include "L0_Platform/test/ram_test.cpp"  // NOLINT
+
+// =============================================================================
 // Interface Test
 // =============================================================================
 #include "L1_Peripheral/test/adc_test.cpp"                // NOLINT
