+++ library/L3_Application/test/unity_test.cpp
@@ -7,7 +7,8 @@
 // =============================================================================
 // FILE I/O
 // =============================================================================
-#include "L3_Application/file_io/test/fatfs_test.cpp"  // NOLINT
+#include "L3_Application/test/fatfs_test.cpp"             // NOLINT
+#include "third_party/fatfs/source/sjsu-dev2/diskio.cpp"  // NOLINT
 
 // =============================================================================
 // Graphics
